+++ src/pages/repair/Repair.jsx
@@ -1,0 +1,5 @@
+import RepairTemplate from '../../templates/repair/RepairTemplate';
+
+export default function Repair() {
+  return <RepairTemplate />;
+}
