--- src/pages/repair/repair-main/Repair.jsx
@@ -1,5 +1,0 @@
-import RepairTemplate from '../../../templates/repair/repair-main/RepairTemplate';
-
-export default function Repair() {
-  return <RepairTemplate />;
-}
